djangoproject - 복사본/requirements.txt
aniso8601==8.0.0
beautifulsoup4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
folium==0.12.1
googlemaps==4.4.5
html5lib
icecream==2.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
JPype1
jsonify==0.5
jupyter==1.0.0
konlpy==0.5.2
lxml
MarkupSafe==1.1.1
mccabe
mysql-connector
mysqlclient
nltk
numpy~=1.23.5
pandas
pandas-datareader==0.9.0
pep8==1.7.1
plotly==5.0.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
seaborn
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.20
tensorflow~=2.11.0
tqdm~=4.64.1
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
torch~=1.13.0
torch
Django~=4.1.4
djangorestframework~=3.14.0
torchvision~=0.14.0
matplotlib~=3.6.2
ipython~=8.7.0
opencv-python~=4.6.0.66
openface~=0.2.1